dev.requirements.txt
-r requirements.txt

black==20.8b1
factory-boy==3.1.0
Faker==4.14.2
flake8==3.8.4
isort==5.6.4
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
